scripts/requirements.txt
paddleocr
Pillow
flask
